[ESP-54] fix build issues, JPEG encoded image saved as . png renamed to .jpg

+++ replace.py
@@ -4,11 +4,13 @@ import shutil
 def get_first_node(path):
     return os.path.split(path)[0].split(os.sep)[0]
 
-one_off_fixes_source = 'fixes'
-one_off_fixes_destination = '..'
+one_off_fixes_source = os.path.join('one-offs', 'fixes')
+one_off_fixes_destination = ''
 
 for root, dirs, files in os.walk(one_off_fixes_source):
     for file in files:
         fix_file = os.path.join(root, file)
-        destination_file = fix_file.replace(get_first_node(fix_file), one_off_fixes_destination)
+        if (not os.path.isfile(fix_file)):
+            continue
+        destination_file = fix_file.replace(one_off_fixes_source + os.path.sep, one_off_fixes_destination)
         shutil.copy(fix_file, destination_file)
